+++ pylinac/__init__.py
@@ -39,6 +39,7 @@ from .planar_imaging import (
     SNCkV,
     StandardImagingFC2,
     IMTLRad,
+    SNCFSQA,
 )
 from .log_analyzer import load_log, Dynalog, TrajectoryLog, MachineLogs
 from .picketfence import PicketFence  # must be after log analyzer